Promenio ponavljanje koda za autorizaciju u statičku funkciju. Uklonio potrebu za autorizacijom za gledanje slika. Manje ispravke.

+++ backend/WebAPI/WebAPI/Controllers/ImageController.cs
@@ -10,6 +10,7 @@ using Microsoft.AspNetCore.Authorization;
 using System.Text.Json;
 using System.IO;
 using Microsoft.Extensions.Hosting.Internal;
+using WebAPI.Helpers;
 
 namespace WebAPI.Controllers
 {
@@ -34,6 +35,9 @@ namespace WebAPI.Controllers
         [Route("get_user_image")]
         public ActionResult<FileData> GetUserImage(uint userId = 0)
         {
+            if (JwtHelper.VerifyToken(currentUserId, Request))
+                return Unauthorized();
+
             if (userId == 0) 
                 userId = currentUserId;
             try
@@ -46,13 +50,11 @@ namespace WebAPI.Controllers
         }
 
 
-        [HttpPost]
-        //[AllowAnonymous]
+        [HttpGet]
+        [AllowAnonymous]
         [Route("get_user_image_file")]
         public IActionResult GetUserImageFile(uint userId = 0)
         {
-            if (userId == 0)
-                userId = currentUserId;
             try
             {
                 string dir = img.GetUserPath(userId);
@@ -76,7 +78,8 @@ namespace WebAPI.Controllers
             }
         }
 
-        [HttpPost]
+        [HttpGet]
+        [AllowAnonymous]
         [Route("get_advert_image_file")]
         public IActionResult GetAdvertImageFile(uint advertId, string imageName)
         {
@@ -97,13 +100,16 @@ namespace WebAPI.Controllers
         [Route("get_advert_image_names")]
         public ActionResult<IEnumerable<string>> GetAdvertImageNames(uint advertId)
         {
+            if (JwtHelper.VerifyToken(currentUserId, Request))
+                return Unauthorized();
+
             try
             {
                 string dir = img.GetAdvertPath(advertId);
                 if (Directory.Exists(dir))
                     return Directory.GetFiles(dir);
                 else
-                    return NotFound();
+                    return Array.Empty<string>();
             }
             catch (Exception e)
             {
@@ -115,7 +121,10 @@ namespace WebAPI.Controllers
         [Route("set_user_image")]
         public ActionResult SetUserImage([FromBody] FileData image)
         {
-            //if (image == null) { return BadRequest("Object must not be null!"); }
+            if (JwtHelper.VerifyToken(currentUserId, Request))
+                return Unauthorized();
+
+            if (image == null) { return BadRequest("Object must not be null!"); }
 
             try
             {
@@ -139,6 +148,9 @@ namespace WebAPI.Controllers
         [Route("delete_user_image")]
         public ActionResult DeleteUserImage()
         {
+            if (JwtHelper.VerifyToken(currentUserId, Request))
+                return Unauthorized();
+
             try
             {
                 img.DeleteUserImage(currentUserId);
@@ -154,6 +166,9 @@ namespace WebAPI.Controllers
         [Route("get_advert_images")]
         public ActionResult<List<FileData>> GetAdvertImages(uint advertId)
         {
+            if (JwtHelper.VerifyToken(currentUserId, Request))
+                return Unauthorized();
+
             try
             {
                 return img.GetAdvertImages(advertId);
@@ -166,6 +181,9 @@ namespace WebAPI.Controllers
 
         private ActionResult TryEditAdvert(uint advertId, Action edit)
         {
+            if (JwtHelper.VerifyToken(currentUserId, Request))
+                return Unauthorized();
+
             if (ctx.Adverts.Where(ad => ad.Id == advertId).FirstOrDefault().OwnerId == currentUserId)
             {
                 try
@@ -186,13 +204,16 @@ namespace WebAPI.Controllers
 
         [HttpPut]
         [Route("add_advert_image")]
-        public ActionResult AddAdvertImage(uint advertId, string image)
+        public ActionResult AddAdvertImage(uint advertId, [FromBody] FileData image)
         {
+            if (JwtHelper.VerifyToken(currentUserId, Request))
+                return Unauthorized();
+
             if (image == null) { return BadRequest("Object must not be null!"); }
             try
             {
-                FileData imageData = JsonSerializer.Deserialize<FileData>(image);
-                return TryEditAdvert(advertId, () => img.AddAdvertImage(advertId, imageData));
+                //FileData imageData = JsonSerializer.Deserialize<FileData>(image);
+                return TryEditAdvert(advertId, () => img.AddAdvertImage(advertId, image));
             }
             catch (Exception e)
             {
@@ -204,6 +225,9 @@ namespace WebAPI.Controllers
         [Route("add_advert_images")]
         public ActionResult AddAdvertImages(uint advertId,[FromBody] List<FileData> images)
         {
+            if (JwtHelper.VerifyToken(currentUserId, Request))
+                return Unauthorized();
+
             //if (images == null) { return BadRequest("Object must not be null!"); }
             try
             {
@@ -221,6 +245,9 @@ namespace WebAPI.Controllers
         [Route("delete_advert_image")]
         public ActionResult DeleteAdvertImage(uint advertId, string imageName = null)
         {
+            if (JwtHelper.VerifyToken(currentUserId, Request))
+                return Unauthorized();
+
             return TryEditAdvert(advertId, () => img.DeleteAdvertImage(advertId, imageName));
         }
 
@@ -228,6 +255,9 @@ namespace WebAPI.Controllers
         [Route("delete_advert_images")]
         public ActionResult DeleteAdvertImages(uint advertId)
         {
+            if (JwtHelper.VerifyToken(currentUserId, Request))
+                return Unauthorized();
+
             return TryEditAdvert(advertId, () => img.DeleteAdvertImage(advertId));
         }
     }
